[Java] Fix issue with putting Long.MIN_VALUE into a buffer via MutableDirectBuffer.putLongAscii(). Issue #145.

+++ agrona/src/main/java/org/agrona/ExpandableArrayBuffer.java
@@ -986,7 +986,7 @@ public class ExpandableArrayBuffer implements MutableDirectBuffer
             return 1;
         }
 
-        if (value == Integer.MIN_VALUE)
+        if (value == Long.MIN_VALUE)
         {
             putBytes(index, MIN_LONG_VALUE);
             return MIN_LONG_VALUE.length;
