+++ agrona/src/main/java/org/agrona/ExpandableDirectByteBuffer.java
@@ -425,7 +425,7 @@ public class ExpandableDirectByteBuffer implements MutableDirectBuffer
             return 1;
         }
 
-        if (value == Integer.MIN_VALUE)
+        if (value == Long.MIN_VALUE)
         {
             putBytes(index, MIN_LONG_VALUE);
             return MIN_LONG_VALUE.length;
